requirements.txt
argon2-cffi==19.1.0
cffi==1.12.3
Django==2.2.5
django-cors-headers==3.1.0
django-debug-toolbar==2.0
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.8
mccabe==0.6.1
mysqlclient==1.4.4
Pillow==6.1.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
